# config/swagger.config.yaml
openapi: 3.0.1
info:
    title: Fashion Shop API
    description: 'This is a Backend API of Fashion Shop'
    termsOfService: https://api.fashionshop.nkt2001.tech/terms/
    contact:
        email: ktuan439@gmail.com
    version: 1.0.0
servers:
    - url: http://localhost:5000/api/v1
    - url: https://api.fashionshop.nkt2001.tech/api/v1
...
paths:
...
    /banner/{id}:
        get:
            summary: Get all banner and slider (non-user, user, admin)
            description: Get all banner and slider (non-user, user, admin)
            tags:
                - Banner
            security:
                - bearerAuth: []
            parameters:
                - name: id
                  in: path
                  description: Banner ID
                  required: true
                  schema:
                      type: string
            responses:
                '200':
                    description: Success
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    success:
                                        type: boolean
                                        default: true
                                    message:
                                        type: string
                                    data:
                                        type: object
                                        properties:
                                            banner:
                                                $ref: '#/components/schemas/Banner'
                '401':
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '404':
                    description: Not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'

        put:
            summary: Update exsited Banner
            tags:
                - Banner
            security:
                - bearerAuth: []
            parameters:
                - name: id
                  in: path
                  description: Banner ID
                  required: true
                  schema:
                      type: string
            requestBody:
                description: Payload to update a new banner
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/BannerForm'
            responses:
                '200':
                    description: Success
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Banner'
                '401':
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '404':
                    description: Not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
        delete:
            summary: Delete existed banner
            description: Delete existed banner
            security:
                - bearerAuth: []
            tags:
                - Banner
            parameters:
                - name: id
                  in: path
                  description: Banner ID
                  required: true
                  schema:
                      type: string
            responses:
                '200':
                    description: Success
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message:
                                        type: string
                '404':
                    description: Not Found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '400':
                    description: Bad request
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '401':
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
        # /banner/{id}/restore:
        #     patch:
        #         summary: Restore disabled Banner
        #         description: Restore disabled Banner
        #         security:
        #             - bearerAuth: []
        #         tags:
        #             - Banner
        #         parameters:
        #             - name: id
        #               in: path
        #               description: Banner ID
        #               required: true
        #               schema:
        #                   type: string
        #         requestBody:
        #             description: isDisabled value
        #             required: true
        #             content:
        #                 application/json:
        #                     schema:
        #                         type: object
        #                         properties:
        #                             isDisabled:
        #                                 type: boolean
        #                                 description: Is disabled or not
        #                                 nullable: false
        #                                 default: false
        #         responses:
        #             '200':
        #                 description: Success
        #                 content:
        #                     application/json:
        #                         schema:
        #                             $ref: '#/components/schemas/Banner'
        #             '404':
        #                 description: Not Found
        #                 content:
        #                     application/json:
        #                         schema:
        #                             $ref: '#/components/schemas/Error'
        #             '401':
        #                 description: Unauthorized
        #                 content:
        #                     application/json:
        #                         schema:
        #                             $ref: '#/components/schemas/Error'
        # /banner/{id}/disable:
        #     patch:
        #         summary: Disable a Banner
        #         description: Disable a Banner
        #         security:
        #             - bearerAuth: []
        #         tags:
        #             - Banner
        #         parameters:
        #             - name: id
        #               in: path
        #               description: Banner ID
        #               required: true
        #               schema:
        #                   type: string
        #         requestBody:
        #             description: isDisabled value
        #             required: true
        #             content:
        #                 application/json:
        #                     schema:
        #                         type: object
        #                         properties:
        #                             isDisabled:
        #                                 type: boolean
        #                                 description: Is disabled or not
        #                                 nullable: false
        #                                 default: true
        #         responses:
        #             '200':
        #                 description: Success
        #                 content:
        #                     application/json:
        #                         schema:
        #                             type: object
        #                             properties:
        #                                 message:
        #                                     type: string
        #             '404':
        #                 description: Not Found
        #                 content:
        #                     application/json:
        #                         schema:
        #                             $ref: '#/components/schemas/Error'
        #             '400':
        #                 description: Bad request
        #                 content:
        #                     application/json:
        #                         schema:
        #                             $ref: '#/components/schemas/Error'
        #             '401':
        #                 description: Unauthorized
        #                 content:
        #                     application/json:
        #                         schema:
        #                             $ref: '#/components/schemas/Error'
...
components:
    schemas:
...
        Banner:
            type: object
            properties:
                _id:
                    type: string
                    description: Banner Id
                title:
                    type: string
                    description: Banner title  or slider title
                index:
                    type: number
                    description: Banner index or slider index
                imageUrl:
                    type: string
                    description: URL image
                linkTo:
                    type: string
                    description: Link to URL
                role:
                    type: string
                    description: Role is banner or slider
                disabled:
                    type: boolean
                    description: Banner or slider is Disabled or not
                    default: false
        BannerForm:
            type: object
            required:
                - title
                - index
                - linkTo
                - role
            properties:
                title:
                    type: string
                    description: Banner title or slider title
                index:
                    type: number
                    description: Banner index or slider index
                # bannerImage:
                #     type: file
                #     format: binary
                #     description: Banner image (select file or fill in imageUrl in the form below)
                imageUrl:
                    type: string
                    description: URL image
                linkTo:
                    type: string
                    description: Link to URL
                role:
                    type: string
                    description: Role is 'banner' or 'slider'
                    default: banner
...
        Error:
            type: object
            properties:
                success:
                    type: boolean
                    default: false
                message:
                    type: string
                    description: error message
                errors:
                    type: array
                    description: error list
                    default: []
                stack:
                    type: string
                    description: error message description
...
    securitySchemes:
        bearerAuth: # arbitrary name for the security scheme
            type: apiKey
            description: "<i>JWT Authorization header using the Bearer scheme. Enter 'Bearer' [space] and then your token in the text input below.</i> <b> Example</b>: <i><u>Bearer 12345abcdef</u></i>"
            in: header # can be "header", "query" or "cookie"
            name: Authorization # name of the header, query parameter or cookie
    # 2) Apply the security globally to all operations
...